Small Task: y2038: Replace timeval with ktime_t
'stLastUpdateTokenAt' is used to store the last updated time
and hence to calculate time taken or difference.

'stLastUpdateTokenAt' is of type 'struct timeval' and according to
y2038 principle, all the instances of 'struct timeval' should be
removed.

'stLastUpdateTokenAt' is converted to 'ktime_t' as ktime_t would be guaranteed
to have sufficient resolution (required resolution: milli seconds).