Staging: dgnc: replace timeval with ktime_t
'timeval' is used to create header timestamps for the data dump.

32-bit systems using 'struct timeval' will break in the year 2038,
So we have to replace that code with more appropriate types.
This patch changes the dgnc driver to use ktime_t instead of timeval.

It uses ktime_get() to get current timestamp. It is better than
do_gettimeofday() as it uses the monotonic clock.